+++ src/shared/Router.jsx
@@ -2,30 +2,31 @@ import React, { useContext } from "react";
 import { BrowserRouter, Routes, Route, Navigate, Outlet } from "react-router-dom";
 import LayOut from "../layout/LayOut";
 import Main from "../pages/Main";
-import Join from "../pages/Join";
 import MyPage from "../pages/MyPage";
 import Board from "../pages/Board";
 import * as S from "../styles/RouterStyle";
 import SignIn from "../components/SignIn";
 import { AggrogramContext } from "../context/AggrogramContext";
+import SignUp from "../components/SignUp";
 
+// 로그인한 상태라면 접근 불가능한 라우터
 const AuthRoute = () => {
   const { user } = useContext(AggrogramContext);
 
-  // if (user) {
-  //   alert("이미 로그인되어 있습니다.");
-  //   return <Navigate to="/" />;
-  // }
+  if (user) {
+    alert("이미 로그인되어 있습니다.");
+    return <Navigate to="/" />;
+  }
   return <Outlet />;
 };
 
 // 로그인한 상태에서만 접근 가능한 라우터
 const PrivateRoute = () => {
   const { user } = useContext(AggrogramContext);
-  // if (!user) {
-  //   alert("로그인이 필요한 페이지입니다. 로그인 페이지로 이동합니다.");
-  //   return <Navigate to="/join" />;
-  // }
+  if (!user) {
+    alert("로그인이 필요한 페이지입니다. 로그인 페이지로 이동합니다.");
+    return <Navigate to="/join" />;
+  }
   return <Outlet />;
 };
 
@@ -39,7 +40,7 @@ const Router = () => {
             <Route path="/board" element={<Board />} />
 
             <Route element={<AuthRoute />}>
-              <Route path="/join" element={<Join />} />
+              <Route path="/signup" element={<SignUp />} />
               <Route path="/signin" element={<SignIn />} />
             </Route>
